Separate App component from index.jsx
