+++ src/Buttons.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-export default function Buttons({ buttonArray, onClick }) {
+export default function Buttons({ numbers, onClick }) {
   return (
     <p>
-      {buttonArray.map((i) => (
+      {numbers.map((i) => (
         <button type="button" key={i} onClick={() => onClick(i)}>
           {i}
         </button>
